feat: allow non-rolling deployments
This adds the ability to do non-rolling deployments, i.e. deployments with
downtime. The default was and is still rolling, so all existing pipelines
should be unaffected.

While Cloud Foundry rolling deployments are for many apps ideal because they
avoid downtime, they don't work with apps that don't have web processes.
data-flow now is made up of several apps, including some that don't need to
have web processes. They do right now, but this is just so they can be deployed
by the existing ci-pipeline code. Having web processes run when they're not
used is a bit wasteful, confusing, and potentially a security issue because
they could be exposed too easily while also forgotten about.

This change means that the web processes of these apps can be scaled to zero,
and still deployed by Jenkins, avoiding any waste and minimising confusion.

When reviewing / a note to the future:

- Keep in mind that we're not Groovy experts. So if there is something odd in
  terms of functionality or style, it is unlikely to be deliberate.
- Hiding whitespace should make it easier to see that the existing code for
  rolling deployments is virtually untouched.

Co-authored-by: Teemerte Teccola <teemerte.teccola@digital.trade.gov.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>